requirements.txt.txt
streamlit==1.12.2
pandas
joblib
sklearn
numpy
pillow

